backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
psycopg2-binary==2.9.1
PyJWT==2.3.0
python-dotenv==0.19.0
Werkzeug==2.0.1 